+++ src/commonMain/kotlin/csense/kotlin/tests/assertions/Arrays.kt
@@ -2,6 +2,7 @@
 
 package csense.kotlin.tests.assertions
 
+import csense.kotlin.annotations.numbers.*
 import kotlin.contracts.ExperimentalContracts
 import kotlin.contracts.InvocationKind
 import kotlin.contracts.contract
@@ -14,7 +15,8 @@ import kotlin.test.assertEquals
  * @param size [Int] the size it should be, if not this will stop the test.
  * @param message [String] the message to print if the assertion fails
  */
-public inline fun IntArray.assertSize(size: Int, message: String = ""): Unit = this.size.assert(size, message)
+public inline fun IntArray.assertSize(@IntLimit(from = 0) size: Int, message: String = ""): Unit =
+    this.size.assert(size, message)
 
 /**
  * Asserts the size of the given [BooleanArray]
@@ -22,7 +24,8 @@ public inline fun IntArray.assertSize(size: Int, message: String = ""): Unit = t
  * @param size [Int] the size it should be, if not this will stop the test.
  * @param message [String] the message to print if the assertion fails
  */
-public inline fun BooleanArray.assertSize(size: Int, message: String = ""): Unit = this.size.assert(size, message)
+public inline fun BooleanArray.assertSize(@IntLimit(from = 0) size: Int, message: String = ""): Unit =
+    this.size.assert(size, message)
 
 /**
  * Asserts the size of the given [DoubleArray]
@@ -30,7 +33,8 @@ public inline fun BooleanArray.assertSize(size: Int, message: String = ""): Unit
  * @param size [Int] the size it should be, if not this will stop the test.
  * @param message [String] the message to print if the assertion fails
  */
-public inline fun DoubleArray.assertSize(size: Int, message: String = ""): Unit = this.size.assert(size, message)
+public inline fun DoubleArray.assertSize(@IntLimit(from = 0) size: Int, message: String = ""): Unit =
+    this.size.assert(size, message)
 
 /**
  * Asserts the size of the given [LongArray]
@@ -38,7 +42,8 @@ public inline fun DoubleArray.assertSize(size: Int, message: String = ""): Unit 
  * @param size [Int] the size it should be, if not this will stop the test.
  * @param message [String] the message to print if the assertion fails
  */
-public inline fun LongArray.assertSize(size: Int, message: String = ""): Unit = this.size.assert(size, message)
+public inline fun LongArray.assertSize(@IntLimit(from = 0) size: Int, message: String = ""): Unit =
+    this.size.assert(size, message)
 
 /**
  * Asserts the size of the given [ByteArray]
@@ -46,7 +51,8 @@ public inline fun LongArray.assertSize(size: Int, message: String = ""): Unit = 
  * @param size [Int] the size it should be, if not this will stop the test.
  * @param message [String] the message to print if the assertion fails
  */
-public inline fun ByteArray.assertSize(size: Int, message: String = ""): Unit = this.size.assert(size, message)
+public inline fun ByteArray.assertSize(@IntLimit(from = 0) size: Int, message: String = ""): Unit =
+    this.size.assert(size, message)
 
 /**
  * Asserts the size of the given [ShortArray]
@@ -54,7 +60,8 @@ public inline fun ByteArray.assertSize(size: Int, message: String = ""): Unit = 
  * @param size [Int] the size it should be, if not this will stop the test.
  * @param message [String] the message to print if the assertion fails
  */
-public inline fun ShortArray.assertSize(size: Int, message: String = ""): Unit = this.size.assert(size, message)
+public inline fun ShortArray.assertSize(@IntLimit(from = 0) size: Int, message: String = ""): Unit =
+    this.size.assert(size, message)
 
 /**
  * Asserts the size of the given [CharArray]
@@ -62,7 +69,8 @@ public inline fun ShortArray.assertSize(size: Int, message: String = ""): Unit =
  * @param size [Int] the size it should be, if not this will stop the test.
  * @param message [String] the message to print if the assertion fails
  */
-public inline fun CharArray.assertSize(size: Int, message: String = ""): Unit = this.size.assert(size, message)
+public inline fun CharArray.assertSize(@IntLimit(from = 0) size: Int, message: String = ""): Unit =
+    this.size.assert(size, message)
 
 /**
  * Asserts the size of the given [FloatArray]
@@ -70,7 +78,8 @@ public inline fun CharArray.assertSize(size: Int, message: String = ""): Unit = 
  * @param size [Int] the size it should be, if not this will stop the test.
  * @param message [String] the message to print if the assertion fails
  */
-public inline fun FloatArray.assertSize(size: Int, message: String = ""): Unit = this.size.assert(size, message)
+public inline fun FloatArray.assertSize(@IntLimit(from = 0) size: Int, message: String = ""): Unit =
+    this.size.assert(size, message)
 
 //region Array<T>
 /**
@@ -79,14 +88,139 @@ public inline fun FloatArray.assertSize(size: Int, message: String = ""): Unit =
  * @param size [Int] the expected size
  * @param message [String]  the message to print if the assertion fails
  */
-public inline fun <T> Array<T>.assertSize(size: Int, message: String = ""): Unit = this.size.assert(size, message)
+public inline fun <T> Array<T>.assertSize(@IntLimit(from = 0) size: Int, message: String = ""): Unit =
+    this.size.assert(size, message)
 
 /**
  * Asserts that this array is empty
  * @receiver [Array]<T>
- * @param message [String]  the message to print if the assertion fails
+ * @param message [String] the message to print if the assertion fails
  */
 public inline fun <T> Array<T>.assertEmpty(message: String = "should be empty"): Unit = assertSize(0, message = message)
+
+/**
+ * Asserts that this array has content
+ * @param message String the assertion message if the array is empty
+ */
+public inline fun <T> Array<T>.assertNotEmpty(message: String = "should have content"): Unit =
+    size.assertLargerOrEqualTo(1, optMessage = message)
+
+
+//region specific array type: assertEmpty
+/**
+ * Asserts that this array is empty
+ * @param message [String] the message to print if the assertion fails
+ */
+public inline fun IntArray.assertEmpty(message: String = "should be empty"): Unit = assertSize(0, message = message)
+
+/**
+ * Asserts that this array is empty
+ * @param message [String] the message to print if the assertion fails
+ */
+public inline fun BooleanArray.assertEmpty(message: String = "should be empty"): Unit = assertSize(0, message = message)
+
+/**
+ * Asserts that this array is empty
+ * @param message [String] the message to print if the assertion fails
+ */
+public inline fun DoubleArray.assertEmpty(message: String = "should be empty"): Unit = assertSize(0, message = message)
+
+/**
+ * Asserts that this array is empty
+ * @param message [String] the message to print if the assertion fails
+ */
+public inline fun LongArray.assertEmpty(message: String = "should be empty"): Unit = assertSize(0, message = message)
+
+/**
+ * Asserts that this array is empty
+ * @param message [String] the message to print if the assertion fails
+ */
+public inline fun FloatArray.assertEmpty(message: String = "should be empty"): Unit = assertSize(0, message = message)
+
+/**
+ * Asserts that this array is empty
+ * @param message [String] the message to print if the assertion fails
+ */
+public inline fun ShortArray.assertEmpty(message: String = "should be empty"): Unit = assertSize(0, message = message)
+
+/**
+ * Asserts that this array is empty
+ * @param message [String] the message to print if the assertion fails
+ */
+public inline fun CharArray.assertEmpty(message: String = "should be empty"): Unit = assertSize(0, message = message)
+//endregion
+
+//region specific array type: assertSingle
+/**
+ * Asserts that this array contains the only given item
+ * @param item Int the item that this array should contain
+ * @param message [String]  the message to print if the equality assertion fails
+ */
+public inline fun IntArray.assertSingle(item: Int, message: String = "") {
+    assertSize(1, message = "Should have 1 item")
+    assertEquals(item, first(), message)
+}
+
+/**
+ * Asserts that this array contains the only given item
+ * @param item Boolean the item that this array should contain
+ * @param message [String]  the message to print if the equality assertion fails
+ */
+public inline fun BooleanArray.assertSingle(item: Boolean, message: String = "") {
+    assertSize(1, message = "Should have 1 item")
+    assertEquals(item, first(), message)
+}
+
+/**
+ * Asserts that this array contains the only given item
+ * @param item Double the item that this array should contain
+ * @param message [String]  the message to print if the equality assertion fails
+ */
+public inline fun DoubleArray.assertSingle(item: Double, message: String = "") {
+    assertSize(1, message = "Should have 1 item")
+    assertEquals(item, first(), message)
+}
+
+/**
+ * Asserts that this array contains the only given item
+ * @param item Long the item that this array should contain
+ * @param message [String]  the message to print if the equality assertion fails
+ */
+public inline fun LongArray.assertSingle(item: Long, message: String = "") {
+    assertSize(1, message = "Should have 1 item")
+    assertEquals(item, first(), message)
+}
+
+/**
+ * Asserts that this array contains the only given item
+ * @param item Float the item that this array should contain
+ * @param message [String]  the message to print if the equality assertion fails
+ */
+public inline fun FloatArray.assertSingle(item: Float, message: String = "") {
+    assertSize(1, message = "Should have 1 item")
+    assertEquals(item, first(), message)
+}
+
+/**
+ * Asserts that this array contains the only given item
+ * @param item Short the item that this array should contain
+ * @param message [String]  the message to print if the equality assertion fails
+ */
+public inline fun ShortArray.assertSingle(item: Short, message: String = "") {
+    assertSize(1, message = "Should have 1 item")
+    assertEquals(item, first(), message)
+}
+
+/**
+ * Asserts that this array contains the only given item
+ * @param item Char the item that this array should contain
+ * @param message [String]  the message to print if the equality assertion fails
+ */
+public inline fun CharArray.assertSingle(item: Char, message: String = "") {
+    assertSize(1, message = "Should have 1 item")
+    assertEquals(item, first(), message)
+}
+//endregion
 
 /**
  * Asserts that this array contains the only given item